More permission optimizing

+++ app/Http/Controllers/Admin/DutyController.php
@@ -6,21 +6,18 @@ use App\Models\Duty;
 use App\Models\Institution;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller as Controller;
+use App\Http\Controllers\ResourceController;
 use App\Models\Role;
 use App\Models\Type;
 use App\Models\User;
+use App\Services\ModelIndexer;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Illuminate\Support\Facades\DB;
 
-class DutyController extends Controller
+class DutyController extends ResourceController
 {
-
-    public function __construct()
-    {
-        $this->authorizeResource(Duty::class, 'duty');
-    }
 
     /**
      * Display a listing of the resource.
@@ -29,19 +26,14 @@ class DutyController extends Controller
      */
     public function index(Request $request)
     {
-        // For search
-        $title = $request->title;
+        $this->authorize('viewAny', [Duty::class, $this->authorizer]);
+        
+        $search = request()->input('title');
 
-        $duties = Duty::when(!is_null($title), function ($query) use ($title) {
-            $query->where('name', 'like', "%{$title}%")->orWhere('email', 'like', "%{$title}%");
-        })->when(!$request->user()->hasRole(config('permission.super_admin_role_name')), function ($query) {
-            $query->whereHas('institution', function ($query) {
-                $query->where('padalinys_id', auth()->user()->padalinys()->id);
-            });
-        })->with(['institution:id,name,short_name,padalinys_id','institution.padalinys:id,shortname'])->paginate(20);
+        $duties = $this->indexer->execute(Duty::class, $search, 'name', $this->authorizer, true);
 
         return Inertia::render('Admin/People/IndexDuty', [
-            'duties' => $duties,
+            'duties' => $duties->with('institution')->paginate(20),
         ]);
     }
 
@@ -52,6 +44,8 @@ class DutyController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', [Duty::class, $this->authorizer]);
+        
         $institutions = $this->getInstitutionsForForm();
         
         return Inertia::render('Admin/People/CreateDuty', [
@@ -68,6 +62,8 @@ class DutyController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', [Duty::class, $this->authorizer]);
+        
         $request->validate([
             'name' => 'required',
             'institution' => 'required',
@@ -96,7 +92,7 @@ class DutyController extends Controller
      */
     public function show(Duty $duty)
     {
-        //
+        $this->authorize('view', [Duty::class, $duty, $this->authorizer]);
     }
 
     /**
@@ -107,6 +103,8 @@ class DutyController extends Controller
      */
     public function edit(Duty $duty)
     {
+        $this->authorize('update', [Duty::class, $duty, $this->authorizer]);
+        
         $institutions = $this->getInstitutionsForForm();
 
         return Inertia::render('Admin/People/EditDuty', [
@@ -131,6 +129,8 @@ class DutyController extends Controller
      */
     public function update(Request $request, Duty $duty)
     {
+        $this->authorize('update', [Duty::class, $duty, $this->authorizer]);
+        
         $request->validate([
             'name' => 'required',
             'institution' => 'required',
@@ -176,6 +176,8 @@ class DutyController extends Controller
      */
     public function destroy(Duty $duty)
     {
+        $this->authorize('delete', [Duty::class, $duty, $this->authorizer]);
+        
         $duty->delete();
 
         return redirect()->route('duties.index')->with('info', 'Pareigybė sėkmingai ištrinta!');
